pod: use websocket exec with spdy fallback
Depends-on: #947

Currently, the pod package always uses the SPDY executor for executing
commands. However, in Kubernetes 1.30, the websocket executor graduated
to beta and is now enabled by default. This PR uses the FallbackExecutor
to default to the websocket executor and only fallback to the SPDY
executor if the websocket executor is unavailable.

Notably, the websocket executor should fix the issue of the stdout being
truncated unexpectedly.